Update README.md
This Python code is designed for analyzing power system faults. It calculates bus currents and visualizes them using a bar chart. In this README, you'll find instructions on how to use the code and explanations of its key components.

•	You can install these libraries using pip to display the figure:

pip install numpy matplotlib
•	To run the code, you can run it in this Visual Studio Code program( I ran the code in this program).
The code will calculate bus currents and display them in the terminal. The bar chart generated by the code will visualize the real and imaginary parts of the currents for each bus. These results to analyse the impact of the fault on the power system.